Moved every single file into subfolders and added all of the styling files, as well as the CenterPiece component and the Greeting component
