+++ chrome/app.html
@@ -1,10 +1,15 @@
 <html>
  <head>
   <title>flow</title>
+  <link href="https://fonts.googleapis.com/css?family=Lato:300,400,700,900" rel="stylesheet">
+  <style>
+  *{
+    font-family: 'Lato', sans-serif;
+    }
+  </style>
  </head>
  <body>
    <div id="app"></div>
    <script src="/bundle.js"></script>
  </body>
 </html>
-
