+++ chrome/src/app.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import { Provider } from 'react-redux';
-import Main from './Main/MainContainer.js';
+import Main from './components/Main/MainContainer.js';
 import store from './store';
 
 ReactDOM.render(
