+++ chrome/src/components/Main/Background/BackgroundVideo/BackgroundVideo.scss
@@ -1,0 +1,8 @@
+.backround-video-wrapper {
+  position: fixed;
+  top: 0;
+  left: 0;
+  z-index: -1;
+  width: 100%;
+  height: 100%;
+}
