+++ chrome/src/components/Main/Main.scss
@@ -1,0 +1,7 @@
+.main {
+  position: fixed;
+  top: 0;
+  left: 0;
+  width: 100vw;
+  height: 100vh;
+}
