+++ chrome/src/components/MainApp/CenterPiece/CenterPiece.scss
@@ -1,0 +1,6 @@
+.center-piece {
+  position: absolute;
+  top: 50%;
+  left: 50%;
+  transform: translate(-50%, -100%);
+}
