+++ chrome/src/components/MainApp/CenterPiece/Clock/Clock.scss
@@ -1,0 +1,9 @@
+.clock {
+  position: relative;
+  color: white;
+  font-size: 34px;
+  font-weight: 300;
+  text-align: center;
+  padding: 18px;
+  z-index: 1;
+}
