+++ chrome/src/components/MainApp/CenterPiece/Greeting/Greeting.jsx
@@ -1,0 +1,16 @@
+import React, { PropTypes } from 'react';
+import './Greeting.scss';
+
+const Greeting = ({
+  name,
+}) => (
+  <div className="greeting">
+    <h1>Hello {name}</h1>
+  </div>
+);
+
+Greeting.propTypes = {
+  name: PropTypes.string,
+};
+
+export default Greeting;
