+++ chrome/src/components/MainApp/CenterPiece/Greeting/Greeting.scss
@@ -1,0 +1,7 @@
+.greeting {
+  font-size: 30px;
+  color: white;
+  h1 {
+    font-weight: 300;
+  }
+}
