+++ chrome/src/components/MainApp/CenterPiece/Greeting/GreetingContainer.js
@@ -1,0 +1,10 @@
+import { connect } from 'react-redux';
+import Greeting from './Greeting.jsx';
+
+const mapStateToProps = ({
+  settings,
+}) => ({
+  name: "Clement",
+});
+
+export default connect(mapStateToProps, null)(Greeting);
