+++ chrome/src/components/MainApp/MainApp.scss
@@ -1,0 +1,5 @@
+.main-app {
+  position: relative;
+  height: 100%;
+  width: 100%;
+}
