+++ chrome/src/components/MainApp/MainAppContainer.js
@@ -1,0 +1,17 @@
+import { connect } from 'react-redux';
+import MainApp from './MainApp.jsx';
+import { setSettings } from '../../reducers/settings';
+
+const mapStateToProps = ({
+  showSettingsPanel,
+}) => ({
+  showSettingsPanel,
+});
+
+const mapDispatchToProps = () => dispatch => ({
+  saveSettings: (settings) => {
+    dispatch(setSettings(settings));
+  },
+});
+
+export default connect(mapStateToProps, mapDispatchToProps)(MainApp);
