+++ chrome/src/components/MainApp/ToggleSettingsPanel/ToggleSettingsPanel.jsx
