+++ chrome/src/components/MainApp/ToggleSettingsPanel/ToggleSettingsPanelContainer.js
@@ -1,0 +1,11 @@
+import { connect } from 'react-redux';
+import { toggleSettingsPanel } from '../../../reducers/showSettingsPanel';
+import ToggleSettingsPanel from './ToggleSettingsPanel.jsx';
+
+const mapDispatchToProps = () => dispatch => ({
+  toggle: () => {
+    dispatch(toggleSettingsPanel());
+  },
+});
+
+export default connect(null, mapDispatchToProps)(ToggleSettingsPanel);
