+++ chrome/src/components/MainApp/Widgets/MeditationWidget/MeditationWidget.jsx
@@ -1,0 +1,15 @@
+import React from 'react';
+
+import { propTypes } from '../../../../reducers/meditationWidget';
+
+const MeditationWidget = ({ meditationWidget }) => (
+  <div className="meditation-widget">
+    MEDITATION WIDGET
+  </div>
+);
+
+MeditationWidget.propTypes = {
+  meditationWidget: propTypes,
+};
+
+export default MeditationWidget;
