+++ chrome/src/components/MainApp/Widgets/WidgetsContainer.js
